outputs/dashboard_project/app/requirements.txt
fastapi
uvicorn
jinja2
psycopg2-binary
python-dotenv
plotly
